feat(suggestions): remove state commands, replaced with respond, implement no update channel updates, make date for suggestions bigint
